"a-z.0cc", an unfinished electronic song
a-z.0cc was a file I made with the chip-tune audio tracker software "dn-famitracker" back in summer-fall 2021. It was gonna be part of 'song pack 2', but it was left unfinished by me due to slowing down progress like tiredness cancels the project; therefore that got abandoned because I wasn't convincing myself enough to complete it or something. The quality and scope of the 'song pack' album series and "a-z" were too demanding back then to me that I had to stop work on both those things entirely. Anyways, I released it here on 2.3.22, because if this xnmcv.top website were finished and published (and yet it remains forever unfinished and unpublished), then the song would've been offered as a download there. I also released the song in its incomplete state because I want this newly revealed song to be displayed to public ears this year, and also for curious historical interest.